[UPDATE] TargetedEffect周りのファイルを分割する
