+++ src/main/java/com/github/unchama/effect/EmptyEffect.kt
@@ -1,0 +1,8 @@
+package com.github.unchama.effect
+
+/**
+ * 何も作用を及ぼさないような[TargetedEffect].
+ */
+object EmptyEffect: TargetedEffect<Any?> {
+  override suspend fun runFor(minecraftObject: Any?) = Unit
+}
