+++ src/main/java/com/github/unchama/effect/TargetedEffect.kt
@@ -1,7 +1,6 @@
 package com.github.unchama.effect
 
 import arrow.typeclasses.Monoid
-import org.bukkit.command.CommandSender
 
 /**
  * Minecraft内の何らかの対象[T]に向けた作用を持ち,
@@ -20,28 +19,3 @@ interface TargetedEffect<in T>{
   }
 }
 
-/**
- * 何も作用を及ぼさないような[TargetedEffect].
- */
-object EmptyEffect: TargetedEffect<Any?> {
-  override suspend fun runFor(minecraftObject: Any?) = Unit
-}
-
-fun String.asMessageEffect() = object : TargetedEffect<CommandSender> {
-  override suspend fun runFor(minecraftObject: CommandSender) {
-    minecraftObject.sendMessage(this@asMessageEffect)
-  }
-}
-
-fun List<String>.asMessageEffect() = object : TargetedEffect<CommandSender> {
-  override suspend fun runFor(minecraftObject: CommandSender) {
-    minecraftObject.sendMessage(toTypedArray())
-  }
-}
-
-operator fun <T> TargetedEffect<T>.plus(anotherEffect: TargetedEffect<T>): TargetedEffect<T> = object : TargetedEffect<T> {
-  override suspend fun runFor(minecraftObject: T) {
-    this@plus.runFor(minecraftObject)
-    anotherEffect.runFor(minecraftObject)
-  }
-}
